Minnor UI tweaks.
