--- YSGBAppleMaps/RotatingButton.swift
@@ -1,21 +1,0 @@
-//
-//  RotatingButton.swift
-//  YSGBAppleMaps
-//
-//  Created by Ярослав on 19.03.2022.
-//
-
-import UIKit
-
-class RotatingButton: UIButton {
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        backgroundColor = UIColor.white.withAlphaComponent(0.95)
-        
-        layer.shadowColor = UIColor.black.cgColor
-        layer.shadowOpacity = 0.25
-        layer.shadowOffset = .zero
-        layer.shadowRadius = 5
-    }
-}
